refactor: change `runs` subcommand tree to `status`
This also renames runs `status` to status `dashboard`
