+++ cmd/subcommands/status.go
@@ -5,20 +5,20 @@ package subcommands
 import (
 	"github.com/spf13/cobra"
 
-	"lunchpail.io/cmd/subcommands/runs"
+	"lunchpail.io/cmd/subcommands/status"
 	"lunchpail.io/pkg/build"
 )
 
 func init() {
 	if build.IsBuilt() {
 		cmd := &cobra.Command{
-			Use:   "runs",
-			Short: "Commands related to runs",
+			Use:   "status",
+			Short: "Commands related to run status",
 		}
 
 		rootCmd.AddCommand(cmd)
-		cmd.AddCommand(runs.Status())
-		cmd.AddCommand(runs.List())
+		cmd.AddCommand(runs.Dashboard())
+		cmd.AddCommand(runs.ListRuns())
 		cmd.AddCommand(runs.Instances())
 		cmd.AddCommand(runs.Cpu())
 	}
